+++ bfe_modules/bfe_modules.go
@@ -19,7 +19,11 @@ package bfe_modules
 import (
 	"github.com/baidu/bfe/bfe_module"
 	"github.com/baidu/bfe/bfe_modules/mod_access"
+	"github.com/baidu/bfe/bfe_modules/mod_auth_basic"
 	"github.com/baidu/bfe/bfe_modules/mod_block"
+	"github.com/baidu/bfe/bfe_modules/mod_compress"
+	"github.com/baidu/bfe/bfe_modules/mod_errors"
+	"github.com/baidu/bfe/bfe_modules/mod_geo"
 	"github.com/baidu/bfe/bfe_modules/mod_header"
 	"github.com/baidu/bfe/bfe_modules/mod_http_code"
 	"github.com/baidu/bfe/bfe_modules/mod_key_log"
@@ -39,23 +43,37 @@ var moduleList = []bfe_module.BfeModule{
 	// Requirement: After mod_trust_clientip
 	mod_logid.NewModuleLogId(),
 
+	// mod_geo
+	// Requirement: After mod_logid
+	mod_geo.NewModuleGeo(),
+
 	// mod_block
-	// Requirement: After mod_dict_client, mod_logid
+	// Requirement: After mod_logid
 	mod_block.NewModuleBlock(),
 
 	// mod_redirect
-	// Requirement: After mod_dict_client, mod_logid
+	// Requirement: After mod_logid
 	mod_redirect.NewModuleRedirect(),
+
+	// mod_auth_basic
+	// Requirement: before mod_static
+	mod_auth_basic.NewModuleAuthBasic(),
+
+	// mod_compress
+	// Requirement: before mod_static
+	mod_compress.NewModuleCompress(),
 
 	// mod_static
 	mod_static.NewModuleStatic(),
 
 	// mod_rewrite
-	// Requirement: After mod_dict_client
 	mod_rewrite.NewModuleReWrite(),
 
 	// mod_header
 	mod_header.NewModuleHeader(),
+
+	// mod_errors
+	mod_errors.NewModuleErrors(),
 
 	// mod_key_log
 	mod_key_log.NewModuleKeyLog(),