+++ bfe_modules/mod_logid/mod_logid.go
@@ -62,23 +62,23 @@ func (m *ModuleLogId) Name() string {
 func (m *ModuleLogId) Init(cbs *bfe_module.BfeCallbacks, whs *web_monitor.WebHandlers,
 	cr string) error {
 	// register handler
-	err := cbs.AddFilter(bfe_module.HANDLE_ACCEPT, m.afterAccept)
+	err := cbs.AddFilter(bfe_module.HandleAccept, m.afterAccept)
 	if err != nil {
 		return fmt.Errorf("%s.Init(): AddFilter(m.afterAccept): %s", m.name, err.Error())
 	}
 
-	err = cbs.AddFilter(bfe_module.HANDLE_BEFORE_LOCATION, m.beforeLocation)
+	err = cbs.AddFilter(bfe_module.HandleBeforeLocation, m.beforeLocation)
 	if err != nil {
 		return fmt.Errorf("%s.Init(): AddFilter(m.beforeLocation): %s", m.name, err.Error())
 	}
 
-	err = cbs.AddFilter(bfe_module.HANDLE_AFTER_LOCATION, m.afterLocation)
+	err = cbs.AddFilter(bfe_module.HandleAfterLocation, m.afterLocation)
 	if err != nil {
 		return fmt.Errorf("%s.Init(): AddFilter(m.afterLocation): %s", m.name, err.Error())
 	}
 
 	// register web handler
-	err = whs.RegisterHandler(web_monitor.WEB_HANDLE_MONITOR, m.name, m.getState)
+	err = whs.RegisterHandler(web_monitor.WebHandleMonitor, m.name, m.getState)
 	if err != nil {
 		return fmt.Errorf("%s.Init(): RegisterHandler(m.getState): %s", m.name, err.Error())
 	}
@@ -89,7 +89,7 @@ func (m *ModuleLogId) Init(cbs *bfe_module.BfeCallbacks, whs *web_monitor.WebHan
 func (m *ModuleLogId) afterAccept(session *bfe_basic.Session) int {
 	session.SessionId = genLogId()
 
-	return bfe_module.BFE_HANDLER_GOON
+	return bfe_module.BfeHandlerGoOn
 }
 
 func (m *ModuleLogId) beforeLocation(req *bfe_basic.Request) (int, *bfe_http.Response) {
@@ -97,7 +97,7 @@ func (m *ModuleLogId) beforeLocation(req *bfe_basic.Request) (int, *bfe_http.Res
 	if req.Session.IsTrustIP {
 		logId := req.HttpRequest.Header.Get(bfe_basic.HeaderBfeLogId)
 		if logId != "" {
-			return bfe_module.BFE_HANDLER_GOON, nil
+			return bfe_module.BfeHandlerGoOn, nil
 		} else {
 			// trust ip, should has a logid
 			m.state.NoLogidFromUpperBfe.Inc(1)
@@ -106,12 +106,12 @@ func (m *ModuleLogId) beforeLocation(req *bfe_basic.Request) (int, *bfe_http.Res
 
 	// generate a new log id
 	req.LogId = genLogId()
-	return bfe_module.BFE_HANDLER_GOON, nil
+	return bfe_module.BfeHandlerGoOn, nil
 }
 
 func (m *ModuleLogId) afterLocation(req *bfe_basic.Request) (int, *bfe_http.Response) {
 	req.HttpRequest.Header.Set(bfe_basic.HeaderBfeLogId, req.LogId)
-	return bfe_module.BFE_HANDLER_GOON, nil
+	return bfe_module.BfeHandlerGoOn, nil
 }
 
 func (m *ModuleLogId) getState(params map[string][]string) ([]byte, error) {