+++ bfe_server/web_server.go
@@ -104,7 +104,7 @@ func (m *BfeMonitor) monitorHandlers() map[string]interface{} {
 		"proxy_handshake_resume_delay": m.srv.proxyHandshakeResumeDelayGet,
 
 		// for module status
-		"module_status": m.srv.ModuleStatusGetJson,
+		"module_status": m.srv.ModuleStatusGetJSON,
 		// for proxy memory stat
 		"proxy_mem_stat": web_monitor.CreateMemStatsHandler("proxy_mem_stat"),
 	}
@@ -132,14 +132,14 @@ func (m *BfeMonitor) reloadHandlers() map[string]interface{} {
 
 func (m *BfeMonitor) WebHandlersInit(srv *BfeServer) error {
 	// register handlers for monitor
-	err := web_monitor.RegisterHandlers(m.WebHandlers, web_monitor.WEB_HANDLE_MONITOR,
+	err := web_monitor.RegisterHandlers(m.WebHandlers, web_monitor.WebHandleMonitor,
 		m.monitorHandlers())
 	if err != nil {
 		return err
 	}
 
 	// register handlers for for reload
-	err = web_monitor.RegisterHandlers(m.WebHandlers, web_monitor.WEB_HANDLE_RELOAD,
+	err = web_monitor.RegisterHandlers(m.WebHandlers, web_monitor.WebHandleReload,
 		m.reloadHandlers())
 	if err != nil {
 		return err